+++ src/plugin.cpp
@@ -22,6 +22,7 @@ void init(Plugin* p) {
 	p->addModel(modelOrbitones);
 	p->addModel(modelAbsorptionSpectrum);
 	p->addModel(modelTalea);
+	p->addModel(modelCollider);
 
 	// Any other plugin initialization may go here.
 	// As an alternative, consider lazy-loading assets and lookup tables when your module is created to reduce startup times of Rack.